Moved Stac properties to StacContext from geoportalContext
